Inital commit
This commit is squashed from multiple earlier commits.

The project started as an experiment to implement a cucumber framework
entirely in Rust with big inspiration from Rocket.
At the start most of the code was copied together from Rocket and other
Cucumber frameworks (mostly cucumber-jvm - the Java implementation).

Because this all started with massive prototyping I did not want to pay
attention to anything that slowed down the progress of getting something
to run:
* No reasonable commit messages
* No comments
* No documentation
* No error handling
* 1-to-1 translations of Java code with all its disadvantages
* No tests
* Bad performance
* ... the list goes on

While most of this stuff does still not exist or is very bad, the basic implementation
seems to work. With this I plan to put more focus on code quality and
will not force push to the repository anymore.

To all the brave readers out there that I still haven't lost:
I hope you have as much fun using the library as I had writing it :)