Enable compilation with stable Rust
To the Rust teams, Rust's contributors,
the entire Rust and Rocket communities,
and Sergio Benitez: thank you all. Sincerely.

Most of the code generation code is **heavily**
based on Rocket (https://rocket.rs/).
Also a lot of the history is shared with it,
including this very commit message.
So now that Rocket compiles on stable Rust
there was no reason that cuke-runner couldn't follow suit.

Thank you all!